default export + bump interfaces dep
