+++ etc/fetch-component.api.md
@@ -1,13 +1,22 @@
-## API Report File for "@well-known-components/base-component"
+## API Report File for "@well-known-components/fetch-component"
 
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
 
 import { IFetchComponent } from '@well-known-components/interfaces';
+import * as nodeFetch from 'node-fetch';
 
 // @public
 export function createFetchComponent(defaultHeaders?: HeadersInit): IFetchComponent;
+
+// @public (undocumented)
+export type RequestOptions = nodeFetch.RequestInit & {
+    abortController?: AbortController;
+    timeout?: number;
+    attempts?: number;
+    retryDelay?: number;
+};
 
 // (No @packageDocumentation comment for this package)
