+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@well-known-components/base-component",
+  "name": "@well-known-components/fetch-component",
   "version": "1.0.0",
-  "description": "base component",
+  "description": "fetch component",
   "main": "dist/index.js",
   "typings": "dist/index.d.ts",
   "scripts": {
@@ -11,13 +11,13 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/well-known-components/base-component.git"
+    "url": "git+https://github.com/well-known-components/fetch-component.git"
   },
   "keywords": [],
   "author": "",
   "license": "Apache-2.0",
   "bugs": {
-    "url": "https://github.com/well-known-components/base-component/issues"
+    "url": "https://github.com/well-known-components/fetch-component/issues"
   },
   "prettier": {
     "printWidth": 120,
@@ -26,7 +26,7 @@
     "trailingComma": "none",
     "tabWidth": 2
   },
-  "homepage": "https://github.com/well-known-components/base-component#readme",
+  "homepage": "https://github.com/well-known-components/fetch-component#readme",
   "devDependencies": {
     "@dcl/eslint-config": "^1.1.1",
     "@microsoft/api-extractor": "^7.34.7",
@@ -40,8 +40,8 @@
     "dist"
   ],
   "dependencies": {
-    "@well-known-components/interfaces": "^1.3.0",
-    "cross-fetch": "^3.1.5",
-    "node-fetch": "^3.3.1"
+    "@types/node-fetch": "^2.5.12",
+    "@well-known-components/interfaces": "^1.4.0",
+    "cross-fetch": "^3.1.5"
   }
 }
