--- src/types.ts
@@ -1,8 +1,0 @@
-import * as nodeFetch from 'node-fetch'
-
-export type RequestOptions = nodeFetch.RequestInit & {
-  abortController?: AbortController
-  timeout?: number
-  attempts?: number
-  retryDelay?: number
-}
